+++ src/server/devices/baseDeviceInterfaces/iHandleSensor.ts
@@ -1,8 +1,8 @@
-import { FensterPosition } from '../models';
+import { WindowPosition } from '../models';
 import { iRoomDevice } from './iRoomDevice';
 
 export interface iHandleSensor extends iRoomDevice {
-  position: FensterPosition;
+  position: WindowPosition;
   minutesOpen: number;
 
   addOffenCallback(pCallback: (pValue: boolean) => void): void;