+++ src/server/devices/hmIPDevices/hmIpRoll.ts
@@ -1,8 +1,8 @@
 import { HmIPDevice } from './hmIpDevice';
 import { DeviceType } from '../deviceType';
 import { LogDebugType, Utils } from '../../services';
-import { Fenster } from '../groups';
-import { FensterPosition } from '../models';
+import { Window } from '../groups';
+import { WindowPosition } from '../models';
 import { LogLevel } from '../../../models';
 import { iShutter } from '../baseDeviceInterfaces';
 import _ from 'lodash';
@@ -35,30 +35,30 @@ export class HmIpRoll extends HmIPDevice implements iShutter {
     if (value !== this._setLevel && Utils.nowMS() - this._setLevelTime < 60 * 10000) {
       value = this._setLevel;
     }
-    if (value !== this._currentLevel && this._fenster) {
+    if (value !== this._currentLevel && this._window) {
       Utils.guardedNewThread(() => {
-        this._fenster?.rolloPositionChange(value);
+        this._window?.rolloPositionChange(value);
       }, this);
       this.persist();
     }
     this._currentLevel = value;
   }
 
-  private _fenster?: Fenster;
+  private _window?: Window;
 
-  public get fenster(): Fenster | undefined {
-    return this._fenster;
+  public get window(): Window | undefined {
+    return this._window;
   }
 
-  public set fenster(value: Fenster | undefined) {
-    this._fenster = value;
+  public set window(value: Window | undefined) {
+    this._window = value;
   }
 
-  public get desiredFensterLevel(): number {
-    if (this._fenster === undefined) {
+  public get desiredWindowShutterLevel(): number {
+    if (this._window === undefined) {
       return -1;
     }
-    return this._fenster.desiredPosition;
+    return this._window.desiredPosition;
   }
 
   public persist(): void {
@@ -102,17 +102,17 @@ export class HmIpRoll extends HmIPDevice implements iShutter {
       return;
     }
 
-    if (this._fenster !== undefined) {
-      if (this._fenster.griffeInPosition(FensterPosition.offen) > 0 && pPosition < 100) {
+    if (this._window !== undefined) {
+      if (this._window.griffeInPosition(WindowPosition.offen) > 0 && pPosition < 100) {
         if (!skipOpenWarning) {
-          this.log(LogLevel.Alert, `Fahre Rollo nicht runter, weil das Fenster offen ist!`);
+          this.log(LogLevel.Alert, `Not closing the shutter, as the window is open!`);
         }
         return;
       }
-      if (this._fenster.griffeInPosition(FensterPosition.kipp) > 0 && pPosition < 50) {
+      if (this._window.griffeInPosition(WindowPosition.kipp) > 0 && pPosition < 50) {
         pPosition = 50;
         if (!skipOpenWarning) {
-          this.log(LogLevel.Alert, `Fahre Rollo nicht runter, weil das Fenster auf Kipp ist!`);
+          this.log(LogLevel.Alert, `Not closing the shutter, as the window is half open!`);
         }
       }
     }
@@ -123,6 +123,6 @@ export class HmIpRoll extends HmIPDevice implements iShutter {
   }
 
   public toJSON(): Partial<IoBrokerBaseDevice> {
-    return _.omit(super.toJSON(), ['_fenster']);
+    return _.omit(super.toJSON(), ['_window']);
   }
 }