+++ src/server/devices/models/WindowPosition.ts
@@ -1,4 +1,4 @@
-export enum FensterPosition {
+export enum WindowPosition {
   geschlossen = 0,
   kipp = 1,
   offen = 2,