+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Rename "Fenster" to window
+
 ## 1.1.5 (2022-09-10)
 
 * (theimo1221) Log More Actuators